+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/application/component/ProductEditForm/TabContent/DataQualityInsights/Attribute.tsx
@@ -1,12 +1,27 @@
-import React, {FunctionComponent} from 'react';
+import React, {FunctionComponent, ReactElement} from 'react';
+import {DATA_QUALITY_INSIGHTS_SHOW_ATTRIBUTE} from "../../../../listener";
 
 interface AttributeProps {
   code: string;
+  label: string;
+  separator: ReactElement | null;
 }
 
-const Attribute: FunctionComponent<AttributeProps> = ({children}) => {
+const handleClick = (attributeCode: string) => {
+  window.dispatchEvent(new CustomEvent(DATA_QUALITY_INSIGHTS_SHOW_ATTRIBUTE, {
+    detail: {
+      code: attributeCode,
+    }
+  }));
+};
+
+const Attribute: FunctionComponent<AttributeProps> = ({code, label, separator}) => {
+
   return (
-    <span className="AknDataQualityInsightsAttribute" data-testid={"dqiAttributeWithRecommendation"}>{children}</span>
+      <button onClick={() => handleClick(code)} className="AknActionButton AknActionButton--withoutBorder AttributeLink">
+        <span className="AknDataQualityInsightsAttribute" data-testid={"dqiAttributeWithRecommendation"}>{label}</span>
+        {separator}
+      </button>
   );
 };