+++ src/Akeneo/Pim/Automation/DataQualityInsights/front/src/infrastructure/reducer/pageContextReducer.ts
@@ -3,10 +3,11 @@ import {ATTRIBUTES_TAB_NAME} from "../../application/constant";
 
 export interface PageContextState {
   currentTab: string;
-  attributesTabIsLoading: boolean
+  attributesTabIsLoading: boolean;
+  attributeToImprove: string | null;
 }
 
-type UpdatePageContextAction = UpdateTabContextAction & UpdateAttributesTabContextAction;
+type UpdatePageContextAction = UpdateTabContextAction & UpdateAttributesTabContextAction & UpdateAttributeToImproveContextAction;
 
 interface UpdateTabContextAction extends Action {
   payload: {
@@ -16,8 +17,14 @@ interface UpdateTabContextAction extends Action {
 
 interface UpdateAttributesTabContextAction extends Action {}
 
+interface UpdateAttributeToImproveContextAction extends Action {
+  payload: {
+    attributeToImprove: string | null;
+  }
+}
+
 export const CHANGE_PRODUCT_TAB = 'CHANGE_PRODUCT_TAB';
-export const changeProductTabAction = (tabName: string): UpdatePageContextAction => {
+export const changeProductTabAction = (tabName: string): UpdateTabContextAction => {
   return {
     type: CHANGE_PRODUCT_TAB,
     payload: {
@@ -40,9 +47,20 @@ export const endProductAttributesTabIsLoadedAction = (): UpdateAttributesTabCont
   };
 };
 
+export const SHOW_DATA_QUALITY_INSIGHTS_ATTRIBUTE_TO_IMPROVE = 'SHOW_DATA_QUALITY_INSIGHTS_ATTRIBUTE_TO_IMPROVE';
+export const showDataQualityInsightsAttributeToImproveAction = (attributeCode: string | null): UpdateAttributeToImproveContextAction => {
+  return {
+    type: SHOW_DATA_QUALITY_INSIGHTS_ATTRIBUTE_TO_IMPROVE,
+    payload: {
+      attributeToImprove : attributeCode
+    }
+  };
+};
+
 const initialState: PageContextState = {
   currentTab: ATTRIBUTES_TAB_NAME,
   attributesTabIsLoading: false,
+  attributeToImprove: null,
 };
 
 const pageContextReducer: Reducer<PageContextState, UpdatePageContextAction> = (previousState = initialState, {type, payload}) => {
@@ -61,6 +79,11 @@ const pageContextReducer: Reducer<PageContextState, UpdatePageContextAction> = (
       return {
         ...previousState,
         attributesTabIsLoading: false
+      };
+    case SHOW_DATA_QUALITY_INSIGHTS_ATTRIBUTE_TO_IMPROVE:
+      return {
+        ...previousState,
+        attributeToImprove: payload.attributeToImprove
       };
     default:
       return previousState;